backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
google-generativeai
python-multipart
aiofiles
passlib[bcrypt]
python-jose[cryptography]